requirements.txt
langchain
google-generativeai 

langchain-community
wikipedia
python-dotenv
openai
streamlit 
gTTs